+++ routing/dht/records.go
@@ -1,32 +1,16 @@
 package dht
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
-	"strings"
 
 	"github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 
 	ci "github.com/jbenet/go-ipfs/p2p/crypto"
-	"github.com/jbenet/go-ipfs/p2p/peer"
+	peer "github.com/jbenet/go-ipfs/p2p/peer"
 	pb "github.com/jbenet/go-ipfs/routing/dht/pb"
-	record "github.com/jbenet/go-ipfs/routing/record"
 	u "github.com/jbenet/go-ipfs/util"
 	ctxutil "github.com/jbenet/go-ipfs/util/ctx"
 )
-
-// ValidatorFunc is a function that is called to validate a given
-// type of DHTRecord.
-type ValidatorFunc func(u.Key, []byte) error
-
-// ErrBadRecord is returned any time a dht record is found to be
-// incorrectly formatted or signed.
-var ErrBadRecord = errors.New("bad dht record")
-
-// ErrInvalidRecordType is returned if a DHTRecord keys prefix
-// is not found in the Validator map of the DHT.
-var ErrInvalidRecordType = errors.New("invalid record keytype")
 
 // KeyForPublicKey returns the key used to retrieve public keys
 // from the dht.
@@ -123,7 +107,7 @@ func (dht *IpfsDHT) verifyRecordLocally(r *pb.Record) error {
 		return fmt.Errorf("do not have public key for %s", p)
 	}
 
-	return dht.verifyRecord(r, pk)
+	return dht.Validator.VerifyRecord(r, pk)
 }
 
 // verifyRecordOnline verifies a record, searching the DHT for the public key
@@ -140,52 +124,5 @@ func (dht *IpfsDHT) verifyRecordOnline(ctx context.Context, r *pb.Record) error 
 		return err
 	}
 
-	return dht.verifyRecord(r, pk)
-}
-
-// TODO: make this an independent exported function.
-// it might be useful for users to have access to.
-func (dht *IpfsDHT) verifyRecord(r *pb.Record, pk ci.PubKey) error {
-	// First, validate the signature
-	blob := record.RecordBlobForSig(r)
-	ok, err := pk.Verify(blob, r.GetSignature())
-	if err != nil {
-		log.Error("Signature verify failed.")
-		return err
-	}
-	if !ok {
-		log.Error("dht found a forged record! (ignored)")
-		return ErrBadRecord
-	}
-
-	// Now, check validity func
-	parts := strings.Split(r.GetKey(), "/")
-	if len(parts) < 3 {
-		log.Infof("Record key does not have validator: %s", u.Key(r.GetKey()))
-		return nil
-	}
-
-	fnc, ok := dht.Validators[parts[1]]
-	if !ok {
-		log.Errorf("Unrecognized key prefix: %s", parts[1])
-		return ErrInvalidRecordType
-	}
-
-	return fnc(u.Key(r.GetKey()), r.GetValue())
-}
-
-// ValidatePublicKeyRecord implements ValidatorFunc and
-// verifies that the passed in record value is the PublicKey
-// that matches the passed in key.
-func ValidatePublicKeyRecord(k u.Key, val []byte) error {
-	keyparts := bytes.Split([]byte(k), []byte("/"))
-	if len(keyparts) < 3 {
-		return errors.New("invalid key")
-	}
-
-	pkh := u.Hash(val)
-	if !bytes.Equal(keyparts[2], pkh) {
-		return errors.New("public key does not match storage key")
-	}
-	return nil
+	return dht.Validator.VerifyRecord(r, pk)
 }
